refactor(gateway): refactor: move Telegram gateway classes to separate package
